Reorganize java integration tests
