+++ .github/workflows/java-integration-tests.yml
@@ -1,0 +1,30 @@
+name: Run Java Integration Tests and Samples
+
+on:
+  push:
+    branches: [ "experimental-java", "java-dev" ]
+  pull_request:
+    branches: [ "experimental-java", "java-dev" ]
+
+jobs:
+  build:
+
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v3
+      - name: Set up JDK 17
+        uses: actions/setup-java@v3
+        with:
+          java-version: '17'
+          distribution: 'temurin'
+          cache: maven
+
+      - name: Build with Maven
+        run: cd java && ./mvnw -B -DskipTests -Pbug-check clean install --file pom.xml
+
+      - name: Run Tests
+        run: cd java/api-test && ../mvnw -B -Pbug-check clean install --file pom.xml
+
+      - name: Build samples
+        run: cd samples/java/semantickernel-samples && ../../../java/mvnw -B clean install --file pom.xml
