+++ java/api-test/api-usage-example/src/main/resources/log4j2.xml
