+++ java/semantickernel-core/src/main/java/com/microsoft/semantickernel/semanticfunctions/DefaultPromptTemplateBuilder.java
@@ -1,9 +1,10 @@
+// Copyright (c) Microsoft. All rights reserved.
 package com.microsoft.semantickernel.semanticfunctions;
 
 public class DefaultPromptTemplateBuilder implements PromptTemplate.Builder {
 
-  @Override
-  public PromptTemplate build(String promptTemplate, PromptTemplateConfig config) {
-    return new DefaultPromptTemplate(promptTemplate, config);
-  }
+    @Override
+    public PromptTemplate build(String promptTemplate, PromptTemplateConfig config) {
+        return new DefaultPromptTemplate(promptTemplate, config);
+    }
 }
